Make some of the animation examples more realistic
Modify the first examples in the animations section so that they reuse
elements of the showcase application at the beginning of the section.
This way we give them a more realistic feel while keeping them simple
enough.

Couldn't be done with all the examples without raising their complexity
too much or diverting from the main point though, so slowly move back to
the simpler examples. It's easy to refer to the showcase for each of
them during explanation though ("remember this animation in the
showcase? it was done exactly with *this* construct").