Refer to the showcase before the end of the states section
Just before the lab in the states/transitions section, refer to the
animations section showcase. This way we force them to think about how
to use states to manage their application workflow. It's the right
opportunity to take a closer look at the showcase application code to
sum up the whole animations/states/transitions triptych.

As a side-effect if should give them some extra clue on how to tackle
the lab just after that.